improved formatting

+++ image_blur/src/main/java/geovista/image_blur/image/MotionBlurOp.java
@@ -97,8 +97,8 @@ public class MotionBlurOp extends AbstractBufferedImageOp {
         if ( dst == null )
             dst = createCompatibleDestImage( src, null );
         BufferedImage tsrc = src;
-        float cx = (float)src.getWidth() * centreX;
-        float cy = (float)src.getHeight() * centreY;
+        float cx = src.getWidth() * centreX;
+        float cy = src.getHeight() * centreY;
         float imageRadius = (float)Math.sqrt( cx*cx + cy*cy );
         float translateX = (float)(distance * Math.cos( angle ));
         float translateY = (float)(distance * -Math.sin( angle ));
